Rename scheduler plugin name
