+++ pkg/scheduler/framework/plugins/registry.go
@@ -1,7 +1,7 @@
 package plugins
 
 import (
-	"github.com/karmada-io/karmada/pkg/scheduler/framework/plugins/apiinstalled"
+	"github.com/karmada-io/karmada/pkg/scheduler/framework/plugins/apienablement"
 	"github.com/karmada-io/karmada/pkg/scheduler/framework/plugins/clusteraffinity"
 	"github.com/karmada-io/karmada/pkg/scheduler/framework/plugins/clusterlocality"
 	"github.com/karmada-io/karmada/pkg/scheduler/framework/plugins/spreadconstraint"
@@ -12,7 +12,7 @@ import (
 // NewInTreeRegistry builds the registry with all the in-tree plugins.
 func NewInTreeRegistry() runtime.Registry {
 	return runtime.Registry{
-		apiinstalled.Name:     apiinstalled.New,
+		apienablement.Name:    apienablement.New,
 		tainttoleration.Name:  tainttoleration.New,
 		clusteraffinity.Name:  clusteraffinity.New,
 		spreadconstraint.Name: spreadconstraint.New,
